"Add reset action and setActiveCollateralAction, fix textButton condition
The code was changed to add reset action at various stages in the manageAaveStateMachine to ensure data consistency. The setActiveCollateralAction was added in the manageCollateral entry to set a default state. In ManageAaveReviewingStateView, the condition for showing textButton is fixed to show only when state matches 'frontend.reviewingAdjusting'."
