+++ features/aave/manage/state/manageAaveStateMachine.ts
@@ -265,7 +265,12 @@ export function createManageAaveStateMachine(
               },
             },
             manageCollateral: {
-              entry: ['reset', 'killCurrentParametersMachine', 'spawnDepositBorrowMachine'],
+              entry: [
+                'reset',
+                'killCurrentParametersMachine',
+                'spawnDepositBorrowMachine',
+                'setActiveCollateralAction',
+              ],
               on: {
                 NEXT_STEP: [
                   {
@@ -571,23 +576,21 @@ export function createManageAaveStateMachine(
         },
         UPDATE_COLLATERAL_TOKEN_ACTION: {
           cond: 'canChangePosition',
-          target: '#manageAaveStateMachine.background.debouncingManage',
-          actions: ['resetTokenActionValue', 'updateCollateralTokenAction'],
+          actions: ['resetTokenActionValue', 'updateCollateralTokenAction', 'reset'],
         },
         UPDATE_CLOSING_ACTION: {
           cond: 'canChangePosition',
           target: '#manageAaveStateMachine.background.debouncingManage',
-          actions: ['resetTokenActionValue', 'updateClosingAction'],
+          actions: ['resetTokenActionValue', 'updateClosingAction', 'reset'],
         },
         UPDATE_DEBT_TOKEN_ACTION: {
           cond: 'canChangePosition',
-          target: '#manageAaveStateMachine.background.debouncingManage',
-          actions: ['resetTokenActionValue', 'updateDebtTokenAction'],
+          actions: ['resetTokenActionValue', 'updateDebtTokenAction', 'reset'],
         },
         UPDATE_TOKEN_ACTION_VALUE: {
           cond: 'canChangePosition',
           target: '#manageAaveStateMachine.background.debouncingManage',
-          actions: ['updateTokenActionValue'],
+          actions: ['updateTokenActionValue', 'reset'],
         },
         USE_SLIPPAGE: {
           target: ['background.debouncing'],
@@ -618,7 +621,8 @@ export function createManageAaveStateMachine(
         isAllowanceNeeded,
         canAdjustPosition: ({ strategyConfig }) =>
           strategyConfig.availableActions().includes('adjust'),
-        isEthersTransaction: ({ strategyConfig }) =>
+        isEthersTransaction: ({ strategyConfig, proxyAddress, transition }) =>
+          allDefined(proxyAddress, transition) &&
           strategyConfig.executeTransactionWith === 'ethers',
       },
       actions: {
@@ -636,6 +640,16 @@ export function createManageAaveStateMachine(
             closingToken,
           },
         })),
+        setActiveCollateralAction: assign(({ manageTokenInput }) => {
+          return {
+            manageTokenInput: {
+              ...manageTokenInput,
+              manageTokenAction:
+                manageTokenInput?.manageTokenAction ??
+                ManageCollateralActionsEnum.DEPOSIT_COLLATERAL,
+            },
+          }
+        }),
         updateCollateralTokenAction: assign(({ manageTokenInput }, { manageTokenAction }) => ({
           manageTokenInput: {
             ...manageTokenInput,
